Splitted manual tests into few files and added test to the main function.
